pyproject.toml
[tool.poetry]
name = "taxdb-poc"
version = "0.1.0"
description = "TaxDB Proof of Concept for Belgium, Spain, and Germany"
authors = ["TaxDB Team"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "3.12.11"
fastapi = "0.111.0"
sqlalchemy = "2.0.29"
pgvector = "0.2.5"
azure-identity = "1.16.0"
azure-storage-blob = "12.19.1"
azure-search-documents = "11.5.0b7"
azure-ai-generative = "1.0.0b1"
pdfminer-six = "20221105"
pytesseract = "0.3.12"
uvicorn = "0.27.0"
python-dotenv = "1.0.0"
tenacity = "^8.2.3"
rich = "^13.7.0"
sentence-transformers = "^2.5.0"
httpx = "0.27.0"
pdfplumber = "^0.10.4"
slowapi = "^0.1.8"

[tool.poetry.group.dev.dependencies]
pytest = "8.2.1"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.5"
docker = "^7.0.0"
psycopg2-binary = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
asyncio_mode = "auto"
markers = [
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests",
    "e2e: marks tests as end-to-end tests",
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/.venv/*", "*/migrations/*", "*/settings.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
    "except ImportError",
]